pyproject.toml
[tool.poetry]
name = "df2 (diploma)"
version = "0.1.0"
description = ""
authors = ["Borodin Dmitriy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
djangorestframework = "^3.15.2"
coverage = "^7.6.8"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
